pyproject.toml
[project]
name = "simple-coding-agent"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.84.0",
    "openai>=2.26.0",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.0",
]
